pyproject.toml
[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 120



[tool.poetry]
name = "chat_fastapi"
version = "0.1.0"
description = "Приложение для общения на FastApi и React"
authors = ["Stanislav Beskaev <stanislav.beskaev@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.76.0"
uvicorn = "^0.17.6"
loguru = "^0.6.0"
SQLAlchemy = "^1.4.36"
coverage = "^6.3.2"
passlib = "^1.7.4"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
bcrypt = "^3.2.2"
python-multipart = "^0.0.5"
pytz = "^2022.1"
requests = "^2.27.1"
starlette-exporter = "^0.14.0"
Jinja2 = "^3.1.2"
websockets = "^10.3"
psycopg2-binary = "2.9.1"
isort = "^5.10.1"

[tool.poetry.dev-dependencies]
pytest = "^7.1.3"
pytest-cov = "^4.0.0"
mypy = "^0.982"
black = "^22.10.0"
pytest-mock = "^3.10.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
